requirements.txt
isort==5.11.4
pycryptodomex==3.16.0
